% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{c2f}
\alias{c2f}
\title{Convert temperature units from C to F}
\usage{
c2f(x)
}
\arguments{
\item{x}{numeric; data in degrees C}
}
\value{
numeric in degrees F
}
\description{
Convert degrees from C to F
}
