% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{calc7dadm}
\alias{calc7dadm}
\title{Calculate the 7dADM for all temperature paths}
\usage{
calc7dadm(xld)
}
\arguments{
\item{xld}{Long-format tibble with Temp-Water data}
}
\value{
Modifies the xld tibble
}
\description{
Tabulate the number of days above or below a threshold
}
