% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{calcDailyMonthlyAnn}
\alias{calcDailyMonthlyAnn}
\title{Calculate Daily, monthly, and Annual values}
\usage{
calcDailyMonthlyAnn()
}
\value{
a list of the original long-format dailys (xld), long-format monthly (xlm), long-format annuals (xla)
}
\description{
Calculate a summary of each variable based on the xw (wide format dataframe)
}
