% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{CalcDatesOfExc}
\alias{CalcDatesOfExc}
\title{Tabulate dates in which a threshold is exceeded}
\usage{
CalcDatesOfExc(DataType, TVals)
}
\arguments{
\item{DataType}{character; column of data to evaluate}

\item{TVals}{numeric; Threshold values in which to tabulate exceedence}
}
\value{
data.frame with name, begin date, end date, and max value
}
\description{
Tabulate dates in which a threshold is exceeded
}
