% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{calcEmergenceTiming}
\alias{calcEmergenceTiming}
\title{Calculate the estimated egg emergence date}
\usage{
calcEmergenceTiming(tout, atu.day = 263, interpMissing = T, hatchValue = 1750)
}
\arguments{
\item{tout}{#tout is a dataframe with the first column as a numeric Julian day, 2nd column temperature data}

\item{interpMissing}{logical; interpolate missing days from previous year/winter if missing?}

\item{hatchValue}{numeric; accumulated thermal units (degree F - days) in which hatching occurs}

\item{spawnDay}{Day of year(s) in which to begin emergence calculation}
}
\value{
data.frame with atu mm/dd, atu day of year, nmiss (missing days)
}
\description{
Tabulate the estimated egg emergence
}
