% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{calcThreshAnn}
\alias{calcThreshAnn}
\title{Tabulate the exceedance values by year for each CWMS data path}
\usage{
calcThreshAnn(xlmExc)
}
\arguments{
\item{xlmExc}{Monthly exceedence tibble (e.g., Temp-Water,TDG,Flow)}
}
\value{
Creates an annuall exceedence tibble
}
\description{
Tabulate the number of days above or below a threshold
}
