% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{calcThreshMon}
\alias{calcThreshMon}
\title{Tabulate the exceedance values by month for each CWMS data path}
\usage{
calcThreshMon(DataType, TVals)
}
\arguments{
\item{DataType}{Type of data (e.g., Temp-Water,TDG,Flow)}

\item{TVals}{Exceedance values to tabulate}
}
\value{
Modifies the xMon tibble
}
\description{
Tabulate the number of days above or below a threshold
}
