% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{CalcTin}
\alias{CalcTin}
\title{Calculate weighted average of inflow temperature}
\usage{
CalcTin(cwms_paths)
}
\value{
modifies the xw (wide format) hourly data for a dam
}
\description{
Calculate the estimated inflow temperature to the reservoir
}
