% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opslogbooks.r
\name{CleanCWMSnames}
\alias{CleanCWMSnames}
\title{Clean up CWMS names, removing unnecessary length of each path name}
\usage{
CleanCWMSnames(x)
}
\arguments{
\item{x}{a data.frame or tibble with column names}
}
\value{
data.frame
}
\description{
Clean up CWMS names, removing unnecessary length of each path name
}
