% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwms_to_dt.r
\name{cwms_to_dt}
\alias{cwms_to_dt}
\title{Function to get CWMS data from one path on the USACE CDA
Written by NBuccola on 2025-01-21 based on Jetilton; taken from https://github.com/jetilton/cwms_read/blob/master/cwms_read.r}
\usage{
cwms_to_dt(
  path,
  start_date,
  end_date,
  timezone = timezone,
  timeseries,
  CDApath = CDApath,
  linux = F
)
}
\arguments{
\item{path}{character; CWMS paths to read data from}

\item{start_date}{character; The begin date of extraction, in 'YYYY-MM-DD HH:MM' format}

\item{end_date}{character; The end date of extraction, in 'YYYY-MM-DD HH:MM' format}

\item{timezone}{character; Example is 'PST8PDT' or 'UTC'}

\item{timeseries}{logical; Use "F" or "FALSE" if retrieving location-level data; some types of location levels not setup yet}

\item{CDApath}{= character; default to 'https://wm.nww.ds.usace.army.mil:8243/nwdp-data/'; untested for other districts...}

\item{linux}{= logical; if running on LINUX OS, use "T"}
}
\value{
a data.frame of the data
}
\description{
Function to get CWMS data from one path on the USACE CDA
Written by NBuccola on 2025-01-21 based on Jetilton; taken from https://github.com/jetilton/cwms_read/blob/master/cwms_read.r
}
\author{
Norman Buccola
}
\keyword{CWMS}
\keyword{Corps}
\keyword{USACE}
\keyword{data}
\keyword{retrieval}
