% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{EstimateEmergence}
\alias{EstimateEmergence}
\title{Tabulate the estimated emergence time for temperature paths}
\usage{
EstimateEmergence(xld, spawnDay = 263, interpMissing = T, hatchValue = 1750)
}
\arguments{
\item{xld}{Daily data tibble with temperature data}

\item{spawnDay}{Day of year(s) in which to begin emergence calculation}

\item{interpMissing}{logical; interpolate missing days from previous year/winter if missing?}

\item{hatchValue}{numeric; accumulated thermal units (degree F - days) in which hatching occurs}
}
\value{
Creates an annuall exceedence tibble
}
\description{
Tabulate the number of days above or below a threshold
}
