% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{f2c}
\alias{f2c}
\title{Convert degrees from F to C}
\usage{
f2c(x)
}
\arguments{
\item{x}{numeric; data in degrees F}
}
\value{
numeric in degrees C
}
\description{
Convert degrees from F to C
}
