% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{filter_complete_years}
\alias{filter_complete_years}
\title{Filter out only years with JDAYs above a threshold}
\usage{
filter_complete_years(df, JDAYthresh)
}
\arguments{
\item{df}{Daily data tibble with temperature data}

\item{JDAYthresh}{What JDAY that is the threshold for keeping years of the data.frame}

\item{date_col}{Which column is the date column}
}
\value{
Creates a trimmed tibble
}
\description{
Filter out only years with adequate JDAYs above a threshold for estimated emergence calcs
}
