% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.r
\name{get_url}
\alias{get_url}
\title{Function to create full URL for cwms_get.r}
\usage{
get_url(
  path,
  start_date,
  end_date,
  timezone = "PST8PDT",
  timeseries = T,
  CDApath = "https://wm.nww.ds.usace.army.mil:8243/nwdp-data/"
)
}
\arguments{
\item{path}{character; CWMS paths to read data from}

\item{start_date}{character; The begin date of extraction, in 'YYYY-MM-DD HH:MM' format}

\item{end_date}{character; The end date of extraction, in 'YYYY-MM-DD HH:MM' format}

\item{timezone}{character; Example is 'PST8PDT' or 'UTC'}

\item{timeseries}{logical; Use "F" or "FALSE" if retrieving location-level data; some types of location levels not setup yet}

\item{CDApath}{= character; default to 'https://wm.nww.ds.usace.army.mil:8243/nwdp-data/'; untested for other districts...}
}
\value{
a data.frame of the data
}
\description{
Function to create full URL for cwms_get.r
}
\author{
Norman Buccola based on Jetilton; https://github.com/jetilton/cwms_read/blob/master/cwms_read.r
}
\keyword{CWMS}
\keyword{Corps}
\keyword{USACE}
\keyword{data}
\keyword{retrieval}
