% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{GetCWMSdata}
\alias{GetCWMSdata}
\title{Get CWMS data from json config file path names}
\usage{
GetCWMSdata(p, bdate, edate, mergeNewPaths = F)
}
\arguments{
\item{p}{list; The site information and pathnames taken from the json config file.}

\item{bdate}{character; The begin date of extraction, in 'YYYY-MM-DD HH:MM' format}

\item{edate}{character; The end date of extraction, in 'YYYY-MM-DD HH:MM' format}
}
\value{
writes raw Rdata files to RdataDir
}
\description{
Use get_cwms to extract data for a time frame for a series of pathnames defined via json config file. See inst
}
