% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opslogbooks.r
\name{getDETBCLScrapedOpsLogBooks}
\alias{getDETBCLScrapedOpsLogBooks}
\title{Get DET Operator Logbook Data}
\usage{
getDETBCLScrapedOpsLogBooks(
  gatePath = file.path(RdataDir, "GateOpenings", "DETBCL"),
  previousScrapeFile = "DETBCL_logbookScraped2014-2025.Rdata",
  newDataFileName
)
}
\arguments{
\item{gatePath}{path to DET gate opening spreadsheets}

\item{previousScrapeFile}{Previous RDATA file that will be merged with}

\item{newDataFileName}{Name of new gate data RDATA file to save}
}
\value{
Saves an Rdata file
}
\description{
Scrapes DET Operator Logbook spreadsheets (monthly)
}
