% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{InterpCols}
\alias{InterpCols}
\title{Interpolate each column's data to hourly}
\usage{
InterpCols(RdataDirProj)
}
\value{
Save data (wide format), project outlet info, pathnames and re-write data to environment (xw)
}
\description{
Interpolate data to hourly and remove columns with all NA values
}
