% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opslogbooks.r
\name{MakeGateOpeningPlots4Will}
\alias{MakeGateOpeningPlots4Will}
\title{Plots of Gate openings and Gate flows per project}
\usage{
MakeGateOpeningPlots4Will(GtDatProj)
}
\arguments{
\item{SaveName}{Name of file to save}
}
\value{
Writes ggplot images
}
\description{
Create Plots of Gate openings and Gate flows per project
}
