% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{MakeMultiYrPlots}
\alias{MakeMultiYrPlots}
\title{Plot the daily averages for each parameter}
\usage{
MakeMultiYrPlots()
}
\arguments{
\item{xld}{daily data}
}
\value{
writes a series of plots based on xl, xld to RdataDir; Writes an annual summary of data (xAnn) to the current environment
}
\description{
Plot daily averages for each parameter
}
