% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opslogbooks.r
\name{MakeProjFlowData4Will}
\alias{MakeProjFlowData4Will}
\title{Create a csv file for Will that has gate-type specific flow at each project for use in thermal plots}
\usage{
MakeProjFlowData4Will(SaveName)
}
\arguments{
\item{SaveName}{Name of file to save}
}
\value{
Writes a csv file for Will that has gate-type specific flow at each project for use in thermal plots
}
\description{
Create a csv file for Will that has gate-type specific flow at each project for use in thermal plots
}
