% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opslogbooks.r
\name{MergeGateOpsWflowWQ}
\alias{MergeGateOpsWflowWQ}
\title{Merge Gate operations data}
\usage{
MergeGateOpsWflowWQ(DETBCLGateOpsFileName, LOPlogbookFileName, GtsCWMS)
}
\arguments{
\item{DETBCLGateOpsFileName}{Scraped DET operating room logbook data from Norm}

\item{LOPlogbookFileName}{Scraped LOP operating room logbook data from Amory}

\item{GtsCWMS}{Gate opening data from CWMS/GDACS}
}
\value{
A tibble with all gate operations data
}
\description{
Merges gate operations data
}
