% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnMetricPlotting.r,
%   R/AnnMetricPlotting_2026-01-30.r
\name{MultiYrQPlots}
\alias{MultiYrQPlots}
\title{Plot the monthly/annual metrics for each type of data}
\usage{
MultiYrQPlots(SubBasinSites, SaveName)

MultiYrQPlots(SubBasinSites, SaveName)
}
\arguments{
\item{SubBasinSites}{data.frame with Proj and subbasin}

\item{SaveName}{The name of the plot}
}
\value{
writes a series of plots based on monthly/annual summaries

writes a series of plots based on monthly/annual summaries
}
\description{
Plot the monthly/annual metrics for each type of data

Plot the monthly/annual metrics for each type of data
}
