% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{pathSummary}
\alias{pathSummary}
\title{Calculate a summary of each variable}
\usage{
pathSummary(x)
}
\arguments{
\item{x}{Assuming a long format dataframe with name = site, value = value}
}
\value{
a summary table of the min, max, quantiles, and percent complete (PrcComp) for each variable.
}
\description{
Calculate a summary of each variable
}
