% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cwms_utilities.r
\name{PlotPercentComplete}
\alias{PlotPercentComplete}
\title{Plot the percentage of complete data by month for each CWMS data path}
\usage{
PlotPercentComplete(xMon)
}
\arguments{
\item{xMon}{Monthly data}
}
\value{
a ggplot object
}
\description{
Plot the percentage of complete data for each CWMS data path
}
