% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_calcs.R
\name{ReformatFishCrit}
\alias{ReformatFishCrit}
\title{Reformat and replicate Fish criteria so it spans a larger timeframe}
\usage{
ReformatFishCrit(FishCrit)
}
\arguments{
\item{FishCrit}{A table of fish criteria with begin/end dates and thresholds}
}
\value{
Long table with Date, name of target, threshold, value
}
\description{
Reformat and replicate Fish criteria so it spans a larger timeframe
}
